Added Volume Support
You may now set the output volume digitally using the added setVol()
function.

It takes one parameter, which is the desired volume in percent, 0-255. A
value of 100 will output the 'Max' volume, which is the normal volume
outputted by the recording.  A value of zero is mute, and a value
greater then 100 will actually boost the original volume.

Careful when boosting, if you go to high, the audio will distort when
the waveform peaks hit the roof and gets clipped off.